DjangoBE/requirement.txt
psycopg2-binary
django
django-widget-tweaks
django-cors-headers
Pillow
django-allauth
django-oauth-toolkit
asgiref
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
python-dotenv